+++ src/cenote/embedders/__init__.py
@@ -3,12 +3,16 @@
 
 from cenote.embedders.base import Embedder
 from cenote.embedders.cache import CachedEmbedder, EmbeddingCache, InMemoryCache
+from cenote.embedders.cohere import CohereEmbedder
 from cenote.embedders.mock import MockEmbedder
+from cenote.embedders.voyage import VoyageEmbedder
 
 __all__ = [
     "CachedEmbedder",
+    "CohereEmbedder",
     "Embedder",
     "EmbeddingCache",
     "InMemoryCache",
     "MockEmbedder",
+    "VoyageEmbedder",
 ]